==> src/multiple_regression/4.2_Residuals+VIF.R <==
##################################################################################
##################################################
###VARIANCE INFLATION FACTOR AND RESIDUAL PLOTS###
##################################################
#
#Takes "processed/input/model/" + crop + "_data.gzip"
#       as generated by 2_data_preprocessing.py 
#and "processed/input/model/" + crop + "_index.csv"
#       as generated by 4.1_GLM_analysis.py
#     with crop = ["Corn", "Rice", "Soybean", "Wheat"]
#as input.
#
#1.) The first part of the code loads the crop data, fits the generalized linear
#     model with gamma distribution and log link and calculates the generalized
#     variance inflation factor for each crop. It saves the GVIF to file.
#
#     Creates the following file:
#       1.) new sheet "Model_VIF" in "reports/Model_results.xlsx"
#2.) The second part creates two types of residual plots for each crop:
#     a) A single scatterplot showing the studentized residuals plotted against
#         the fitted values on the link scale.
#     b) Four plots based on the standardized deviance residuals and the fitted
#         values on the response scale: Residuals vs. Fitted, Q-QPlot,
#         Location-Scale Plot, Residual-Leverage Plot.
#
#     Creates the following files:
#       1.) "reports/figures/" + crop + "_residuals_link.png"
#       2.) "reports/figures/" + crop + "_residuals.png"
#           with crop = ["Corn", "Rice", "Soybean", "Wheat"]
#
#@author: Jessica Mörsdorf
#jessica@allfed.info
#jessi.moersdorf@gmail.com
#################################################################################

# Install and load necessary packages
install.packages("ggResidpanel")
install.packages("openxlsx")
install.packages("emmeans")
library(car)
library(openxlsx)
library(ggplot2)
library("ggResidpanel")
library(dplyr)
library(emmeans)

# Define input file path
getwd()
file_path <- "processed/input/model/"

# Specify the crops
crops <- c("Corn", "Rice", "Soybean", "Wheat")

#######################################################
###Load and prepare model input data###################
######################################################

print("Reading crop and index data and defining calibration data")

# Load the clean data set for each crop and extract the model calibration
# data set by dropping all rows of the validation data set

model_data <- list()
val_index <- list()
fit_data <- list()

for (crop in crops) {
  model_data[[crop]] <- read.csv(paste0(file_path, crop, "_data.gzip"), 
                                 header=TRUE, sep=",", row.names=1, quote="", 
                                 comment="", stringsAsFactors=FALSE)
  val_index[[crop]] <- read.csv(paste0(file_path, crop, "_index.csv"), 
                                 header=TRUE, sep=",", quote="", 
                                 comment="", stringsAsFactors=FALSE)
  model_data[[crop]][, sapply(model_data[[crop]], is.integer)] <- lapply(model_data[[crop]][,
                                          sapply(model_data[[crop]], is.integer)], as.factor)

  #subtract the rows of the val_index data (which represent 20% of the model data,
  #to be used for model validation) from the clean data set to be left with 80%
  #of the model data to fit the model
  fit_data[[crop]] <- model_data[[crop]][!rownames(model_data[[crop]]) %in% val_index[[crop]][["X0"]],]
}

print("Done reading crop and index data and defining calibration data")


###########################################################
### Fit gamma GLM, calculate VIF and save it to file#######
###########################################################

print("Fitting the GLM and calculating the VIF")

model <- list()
VIF <- list()
inf_fact <- list()

for (crop in crops) {
  
  # Fit gamma GLM with log link
  
  #specify the model for each crop with a gamma distribution and the independent factors total nitrogen application,
  #artificial phosphorus fertilizer application, total irrigation, pesticide application, temperature + 
  #moisture + soil class and a dummy indicating if the area is worked with agricultural machinery
  model[[crop]] <- glm(
  formula = Yield ~ n_total + irrigation_tot + mechanized + pesticides + 
    thz_class + mst_class + soil_class,
  data = fit_data[[crop]],
  family = Gamma(link = "log")
  )

  #Calculate VIF
  
  #calculate GVIF, Df and GVIF^1/2*Df and save it in a data frame
  inf_fact[[crop]] <- vif(model[[crop]])
  VIF[[crop]] <- data.frame(inf_fact[crop])
}

#combine all four data frames of VIF[[crop]] into one data frame
VIF_crops <- bind_cols(VIF)

#create a new column for each crop that squares the "GVIF^1/2*Df" column
#for variables with Df=1 the squared "GVIF^1/2*Df" values equal the GVIF
#this new column can be interpreted with the same rules of thumb as the VIF
for (crop in crops) {
  VIF_crops <- VIF_crops %>% 
    mutate(!!paste0(crop, ".GVIF2") := (!!sym(paste0(crop, ".GVIF..1..2.Df.."))) ^ 2)
}

#order the columns alphabetically
VIF_crops <- VIF_crops[, order(names(VIF_crops))]
VIF_crops <- round(VIF_crops, digits=2)

print("Done fitting the GLM and calculating the VIF")


print("Saving the VIF to the 'reports/Model_Results.xlsx' file")

# Save the resulting VIF data frame to an already existing Excel file

#create a style for the column names
hs2 <- createStyle(
  fontColour = "#000000", halign = "center",
  valign = "center", textDecoration = "bold",
  border = "TopBottomLeftRight"
)
# Load the existing Excel file
wb <- loadWorkbook("reports/Model_results.xlsx")

# remove the worksheet if a sheet with that name already exists
#removeWorksheet(wb, sheet = "Model_VIF")
# Add a new sheet with your data frame
addWorksheet(wb, "Model_VIF")
writeDataTable(wb, sheet = "Model_VIF", x = VIF_crops, rowNames=TRUE,
               tableStyle = "None", headerStyle = hs2, firstColumn = TRUE)

# Save the updated Excel file
saveWorkbook(wb, "reports/Model_results.xlsx", overwrite = TRUE)

print("Done saving the VIF to the 'reports/Model_Results.xlsx' file")


##################################################
### Create Residual Plots and save them to file###
##################################################

print("Creating the plot of studentized residuals vs. fitted values on link scale/
      and saving it to file")

#specify the path where the plots will be saved
file_path2 <- "reports/figures/"

# Create a residual plot of studentized residuals vs. fitted values
# on the link scale

for (crop in crops){
  
  #Combine the residuals and fitted values into a data frame
  resid_df <- data.frame(predict(model[[crop]], type="link"), rstudent(model[[crop]]))
  
  #rename the columns
  colnames(resid_df) <- setNames(c("fitted_link", "stud_resid"),
                                 c("fitted_link", "stud_resid"))
  
  #create a scatterplot and assign it to a variable
  resid_plot <- ggplot(data = resid_df, aes(x = fitted_link, y = stud_resid)) +
    geom_point(alpha = 0.1, shape=1) +
    geom_hline(yintercept = 0, linetype = "dashed")
  
  #save the scatterplot to file
  ggsave(paste0(file_path2, crop, "_residuals_link.png"),plot = resid_plot,
         width = 6, height = 4, dpi = 300)
}

print("Done creating the plot of studentized residuals vs. fitted values on link scale/
      and saving it to file")


# Create a panel of four residual plots based on the standardized deviance
# residuals and the fitted values on the response scale

print("Creating the panel of four residual plots based on the standardized deviance
      residuals and saving it to file")

for (crop in crops){
  
  #create a panel with four plots: resid vs. fitted, Q-QPlot, Location-Scale Plot,
  #Residual-Leverage Plot and assign it to a variable
  resid_pan <- resid_panel(model[[crop]], plots="R", type="stand.deviance",
              theme="grey", axis.text.size = 8)
  
  #save the residual panel to file
  ggsave(paste0(file_path2, crop, "_residuals.png"),plot = resid_pan,
         width = 6, height = 4, dpi = 300)
}

print("Done creating the panel of four residual plots based on the standardized deviance
      residuals and saving it to file")
